requirements.txt
# p1

stable_baselines3
pyyaml
gymnasium
numpy
robobopy
robobosim
scipy

# p2

neat-python
graphviz
tqdm
